[#155206093] controller refactored
